+++ src/components/code/utils.tsx
@@ -15,12 +15,21 @@ type ExtendedRefractorNode = RefractorNode & {
   lineEnd?: number;
 };
 
+interface LineNumbersConfig {
+  start: number;
+  show: boolean;
+}
+
+// Approximate width of a single digit/character
+const CHAR_SIZE = 8;
+const $euiSizeS = 8;
+
 const isAstElement = (node: RefractorNode): node is AST.Element =>
   node.hasOwnProperty('type') && node.type === 'element';
 
 const addLineData = (
   nodes: ExtendedRefractorNode[],
-  data = { lineNumber: 1 }
+  data: { lineNumber: number }
 ): ExtendedRefractorNode[] => {
   return nodes.reduce<ExtendedRefractorNode[]>((result, node) => {
     const lineStart = data.lineNumber;
@@ -68,7 +77,10 @@ const addLineData = (
   }, []);
 };
 
-function wrapLines(nodes: ExtendedRefractorNode[]) {
+function wrapLines(
+  nodes: ExtendedRefractorNode[],
+  options: { showLineNumbers: boolean }
+) {
   const grouped: ExtendedRefractorNode[][] = [];
   nodes.forEach((node) => {
     const lineStart = node.lineStart! - 1;
@@ -79,14 +91,40 @@ function wrapLines(nodes: ExtendedRefractorNode[]) {
     }
   });
   const wrapped: RefractorNode[] = [];
-  grouped.forEach((node) => {
+  const digits = grouped.length.toString().length;
+  const width = digits * CHAR_SIZE;
+  grouped.forEach((node, i) => {
+    const children: RefractorNode[] = options.showLineNumbers
+      ? [
+          {
+            type: 'element',
+            tagName: 'span',
+            properties: {
+              style: { width },
+              ['data-line-number']: i + 1,
+              ['aria-hidden']: true,
+              className: ['euiCodeBlock__lineNumber'],
+            },
+            children: [],
+          },
+          {
+            type: 'element',
+            tagName: 'span',
+            properties: {
+              style: { marginLeft: width + $euiSizeS },
+              className: ['euiCodeBlock__lineText'],
+            },
+            children: node,
+          },
+        ]
+      : node;
     wrapped.push({
       type: 'element',
       tagName: 'span',
       properties: {
         className: ['euiCodeBlock__line'],
       },
-      children: node,
+      children,
     });
   });
   return wrapped;
@@ -117,6 +155,13 @@ export const nodeToHtml = (
   return <React.Fragment key={key}>{node.value}</React.Fragment>;
 };
 
-export const highlightByLine = (children: string, language: string) => {
-  return wrapLines(addLineData(highlight(children, language)));
+export const highlightByLine = (
+  children: string,
+  language: string,
+  data: LineNumbersConfig
+) => {
+  return wrapLines(
+    addLineData(highlight(children, language), { lineNumber: data.start }),
+    { showLineNumbers: data.show }
+  );
 };